[+]更新Todo相关功能和界面 [fixed]修复了对Todo异常数据的处理

+++ app/src/main/java/watermelon/tobe/viewmodel/DayFragmentViewModel.kt
@@ -3,6 +3,7 @@ package watermelon.tobe.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.collectLatest
@@ -20,7 +21,7 @@ class DayFragmentViewModel : ViewModel() {
     val todoList = MutableStateFlow<List<Todo>>(listOf())
     fun queryTodoList(date: String) {
         viewModelScope.launch(Dispatchers.IO) {
-            TodoRepository.queryTodoList(date = date).collectLatest {
+            TodoRepository.queryTodoList(date = date, status = 0).collectLatest {
                 todoList.emit(it)
             }
         }
@@ -28,12 +29,34 @@ class DayFragmentViewModel : ViewModel() {
 
     fun deleteTodo(todo: Todo) {
         viewModelScope.launch(Dispatchers.IO) {
-            TodoRepository.deleteTodo(todo)
-            todoList.update {
-                for (i in 0..it.size){
-                    if (it[i] == todo) return@update it.subList(0,i)+it.subList(i+1,it.size)
+            if (todo.priority!=-1){//用if排除Empty,Connect Fail,Loading状态的T odo
+                TodoRepository.deleteTodo(todo)
+                todoList.update {
+                    for (i in 0..it.size) {
+                        if (it[i] == todo) return@update it.subList(0, i) + it.subList(i + 1, it.size)
+                    }
+                    return@update it
                 }
-                return@update it
+            }
+        }
+    }
+
+    //完成todo
+    fun finishTodo(todo: Todo) {
+        viewModelScope.launch(Dispatchers.IO) {
+            if (todo.priority!=-1){//用if排除Empty,Connect Fail,Loading状态的T odo
+                TodoRepository.updateTodo(
+                    id = todo.id,
+                    title = todo.title,
+                    content = todo.content,
+                    dateStr = todo.dateStr,
+                    priority = todo.priority,
+                    status = 1
+                )
+                delay(10)
+                TodoRepository.queryTodoList(date = todo.dateStr, status = 0).collectLatest {
+                    todoList.emit(it)
+                }
             }
         }
     }
